+++ src/Utility/BackupImport.php
@@ -23,8 +23,6 @@
  */
 namespace MysqlBackup\Utility;
 
-use Cake\Core\Configure;
-use Cake\Datasource\ConnectionManager;
 use Cake\Network\Exception\InternalErrorException;
 use MysqlBackup\BackupTrait;
 
@@ -62,7 +60,7 @@ class BackupImport
      */
     public function __construct()
     {
-        $this->connection = ConnectionManager::getConfig(Configure::read(MYSQL_BACKUP . '.connection'));
+        $this->connection = $this->getConnection();
 
         $driver = (new \ReflectionClass($this->connection['driver']))->getShortName();
         $driver = MYSQL_BACKUP . '\\Driver\\' . $driver;