+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # 1.x branch
 ## 1.1 branch
 ### 1.1.1
+* `_storeAuth()` methods from `BackupExport` and `BackupImport` classes are
+    now private;
+* added `BackupTrait::getConnection()` method;
 * added `BackupTrait::getValidCompressions()` method.
 
 ### 1.1.0